back/requirements.txt
Django==5.0.4
pillow==10.3.0
pillow_heif==0.16.0
psycopg2-binary==2.9.9
wcwidth==0.2.13
python-dotenv==1.0.1
gunicorn==22.0.0
redis==5.0.4
djangorestframework==3.15.1
djangorestframework-datatables==0.7.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
djoser==2.2.3
openpyxl==3.1.2
